% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promedio_temperaturas.R
\name{promedio_temperaturas}
\alias{promedio_temperaturas}
\title{Calcular el promedio de dos temperaturas}
\usage{
promedio_temperaturas(temp1, temp2)
}
\arguments{
\item{temp1}{Primera temperatura.}

\item{temp2}{Segunda temperatura.}
}
\value{
El promedio de las dos temperaturas.
}
\description{
Calcula el promedio de dos valores de temperatura.
}
\examples{
promedio_temperaturas(20, 30)
}
