% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion_prueba.R
\name{suma}
\alias{suma}
\title{Sumar dos numeros}
\usage{
suma(x = 0, y = 0)
}
\arguments{
\item{x}{Primer numero. Por defecto es 0.}

\item{y}{Segundo numero. Por defecto es 0.}
}
\value{
La suma de los numeros o un mensaje
si alguno es negativo.
}
\description{
Suma dos valores numericos no negativos.
}
\examples{
suma(3, 4)
suma()
suma(-2, 5)
}
