% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos.R
\docType{data}
\name{temperaturas}
\alias{temperaturas}
\title{Temperaturas de ejemplo}
\format{
Vector numerico con 5 valores.
}
\source{
Datos ficticios creados para Programacion II.
}
\usage{
temperaturas
}
\description{
Conjunto de cinco temperaturas utilizado para
practicar el manejo de datos en el paquete meteoDatos.
}
\keyword{datasets}
